+++ lib/src/models/mpt_prayer.g.dart
@@ -9,9 +9,15 @@ part of 'mpt_prayer.dart';
 MptPrayer _$MptPrayerFromJson(Map<String, dynamic> json) => MptPrayer(
   day: (json['day'] as num).toInt(),
   hijri: HijriDate.fromJson(json['hijri'] as String),
+  imsak: const EpochDateTimeConverter().fromJson(
+    (json['imsak'] as num).toInt(),
+  ),
   fajr: const EpochDateTimeConverter().fromJson((json['fajr'] as num).toInt()),
   syuruk: const EpochDateTimeConverter().fromJson(
     (json['syuruk'] as num).toInt(),
+  ),
+  dhuha: const EpochDateTimeConverter().fromJson(
+    (json['dhuha'] as num).toInt(),
   ),
   dhuhr: const EpochDateTimeConverter().fromJson(
     (json['dhuhr'] as num).toInt(),
@@ -26,8 +32,10 @@ MptPrayer _$MptPrayerFromJson(Map<String, dynamic> json) => MptPrayer(
 Map<String, dynamic> _$MptPrayerToJson(MptPrayer instance) => <String, dynamic>{
   'day': instance.day,
   'hijri': instance.hijri,
+  'imsak': const EpochDateTimeConverter().toJson(instance.imsak),
   'fajr': const EpochDateTimeConverter().toJson(instance.fajr),
   'syuruk': const EpochDateTimeConverter().toJson(instance.syuruk),
+  'dhuha': const EpochDateTimeConverter().toJson(instance.dhuha),
   'dhuhr': const EpochDateTimeConverter().toJson(instance.dhuhr),
   'asr': const EpochDateTimeConverter().toJson(instance.asr),
   'maghrib': const EpochDateTimeConverter().toJson(instance.maghrib),