+++ lib/src/models/mpt_prayer_time.dart
@@ -1,14 +1,18 @@
 import 'package:json_annotation/json_annotation.dart';
 
+import 'hijri_date.dart';
+
 part 'mpt_prayer_time.g.dart';
 
 @JsonSerializable()
 class MptPrayerTime {
-  final String hijri;
+  final HijriDate hijri;
   final String date;
   final String day;
+  final String imsak;
   final String fajr;
   final String syuruk;
+  final String dhuha;
   final String dhuhr;
   final String asr;
   final String maghrib;
@@ -18,8 +22,10 @@ class MptPrayerTime {
     required this.hijri,
     required this.date,
     required this.day,
+    required this.imsak,
     required this.fajr,
     required this.syuruk,
+    required this.dhuha,
     required this.dhuhr,
     required this.asr,
     required this.maghrib,