+++ lib/src/models/mpt_prayer_time.g.dart
@@ -8,11 +8,13 @@ part of 'mpt_prayer_time.dart';
 
 MptPrayerTime _$MptPrayerTimeFromJson(Map<String, dynamic> json) =>
     MptPrayerTime(
-      hijri: json['hijri'] as String,
+      hijri: HijriDate.fromJson(json['hijri'] as String),
       date: json['date'] as String,
       day: json['day'] as String,
+      imsak: json['imsak'] as String,
       fajr: json['fajr'] as String,
       syuruk: json['syuruk'] as String,
+      dhuha: json['dhuha'] as String,
       dhuhr: json['dhuhr'] as String,
       asr: json['asr'] as String,
       maghrib: json['maghrib'] as String,
@@ -24,8 +26,10 @@ Map<String, dynamic> _$MptPrayerTimeToJson(MptPrayerTime instance) =>
       'hijri': instance.hijri,
       'date': instance.date,
       'day': instance.day,
+      'imsak': instance.imsak,
       'fajr': instance.fajr,
       'syuruk': instance.syuruk,
+      'dhuha': instance.dhuha,
       'dhuhr': instance.dhuhr,
       'asr': instance.asr,
       'maghrib': instance.maghrib,